+++ src/acceleration/bvh3.hpp
@@ -26,26 +26,26 @@ public:
 
     // Hittable
     bool
-    hit(const Ray& r, double tMin, double tMax, HitRecord& rec) const override
+    hit(const Ray& r, float tMin, float tMax, HitRecord& rec) const override
     {
         return intersectIterative(rootNodeIdx, r, tMin, tMax, rec);
     }
 
-    std::string tree(size_t nodeIdx, int depth = 0) const;
+    std::string tree(uint32_t nodeIdx, int depth = 0) const;
 
-    size_t getNodesUsed() const { return nodesUsed - 1; }
+    uint32_t getNodesUsed() const { return nodesUsed - 1; }
 
     struct Node {
         Aabb aabb;
-        size_t primCount; ///< For a leaf: primitive count, for an inner node: 0
+        uint32_t primCount; ///< For a leaf: primitive count, for an inner node: 0
         union {
-            size_t mLeftChildIdx; ///< Index of left child node, for inner nodes
-            size_t firstPrimIdx;  ///< Index of first primitive, for leaf nodes
+            uint32_t mLeftChildIdx; ///< Index of left child node, for inner nodes
+            uint32_t firstPrimIdx;  ///< Index of first primitive, for leaf nodes
         };
 
         bool isLeaf() const { return primCount > 0; }
-        size_t left() const { return mLeftChildIdx; }
-        size_t right() const { return mLeftChildIdx + 1; }
+        uint32_t left() const { return mLeftChildIdx; }
+        uint32_t right() const { return mLeftChildIdx + 1; }
         float cost() const { return primCount * aabb.area(); }
     };
 
@@ -57,24 +57,24 @@ public:
 private:
     /// @brief Update AABB bounds of root node.
     /// @param nodeIdx
-    void updateNodeBounds(const size_t nodeIdx);
+    void updateNodeBounds(const uint32_t nodeIdx);
 
-    double findBestSplitPlane(Node& node, int& axis, double& splitPos);
+    float findBestSplitPlane(Node& node, int& axis, float& splitPos);
 
     /// @brief Recursive BVH building. Determine split axis and position, split
     /// and partition primitives, create child nodes and recurse.
     /// @param nodeIdx
-    void subdivide(const size_t nodeIdx);
+    void subdivide(const uint32_t nodeIdx);
 
     /// @brief Recursive intersection function called by hit(). First check if
     /// AABB intersects ray, in which case either; For leaf nodes, compute hit
     /// similarly to HittableList (exactly what this is); For inner nodes,
     /// recurse and handle results.
     bool intersect(
-        const size_t nodeIdx,
+        const uint32_t nodeIdx,
         const Ray& r,
-        const double tMin,
-        const double tMax,
+        const float tMin,
+        const float tMax,
         HitRecord& rec,
         int depth = 0) const;
 
@@ -83,10 +83,10 @@ private:
     /// similarly to HittableList (exactly what this is); For inner nodes,
     /// recurse and handle results.
     bool intersectIterative(
-        const size_t nodeIdx,
+        const uint32_t nodeIdx,
         const Ray& r,
-        const double tMin,
-        const double tMax,
+        const float tMin,
+        const float tMax,
         HitRecord& rec) const;
 
     /// @brief Evaluate SAH for the node on the given split axis and position.
@@ -96,8 +96,8 @@ private:
     float evaluateSAH(Node& node, int axis, float pos);
 
 private:
-    size_t rootNodeIdx = 0;
-    size_t nodesUsed   = 2;
+    uint32_t rootNodeIdx = 0;
+    uint32_t nodesUsed   = 2;
     std::vector<Node> nodes;
     /// @brief Reference to list of primitives. Assume that this one can be
     /// shared among subsystems, as so should not be modified.
@@ -107,6 +107,6 @@ private:
     /// even 2^32).
     std::vector<uint32_t> primIndices;
     /// Primitives size
-    size_t N;
+    uint32_t N;
 };
 } // namespace blikker_basic
