+++ src/shape/quad.hpp
@@ -18,7 +18,7 @@ public:
     virtual ~Quad() = default;
 
     bool
-    hit(const Ray& ray, double tMin, double tMax, HitRecord& rec) const override
+    hit(const Ray& ray, float tMin, float tMax, HitRecord& rec) const override
     {
         quadIntersections++;
         // https://raytracing.github.io/books/RayTracingTheNextWeek.html
@@ -54,7 +54,7 @@ public:
         return true;
     }
 
-    Vec3 centroid() const override { return q + (uEdge + vEdge) / 2.0; }
+    Vec3 centroid() const override { return q + (uEdge + vEdge) / 2.0f; }
     void growAABB(Aabb& aabb) const override
     {
         aabb.grow(q);
@@ -71,7 +71,7 @@ private:
     /// @param b
     /// @param rec
     /// @return
-    bool isInterior(double a, double b, HitRecord& rec) const
+    bool isInterior(float a, float b, HitRecord& rec) const
     {
         if ((a < 0) || (1 < a) || (b < 0) || (1 < b)) return false;
