+++ src/shape/sphere.hpp
@@ -10,16 +10,16 @@
 class Sphere : public Primitive {
 public:
     Sphere() { }
-    Sphere(Vec3 c, double r, shared_ptr<Material> m)
+    Sphere(Vec3 c, float r, shared_ptr<Material> m)
         : center(c)
         , radius(r)
         , mat(m)
     { }
     virtual ~Sphere() = default;
 
-    void getSphereUV(const Vec3& p, double& u, double& v) const;
+    void getSphereUV(const Vec3& p, float& u, float& v) const;
 
-    virtual bool hit(const Ray& ray, double tMin, double tMax, HitRecord& rec)
+    virtual bool hit(const Ray& ray, float tMin, float tMax, HitRecord& rec)
         const override;
     virtual Vec3 centroid() const override { return center; }
     virtual void growAABB(Aabb& aabb) const override {
@@ -29,7 +29,7 @@ public:
 
 public:
     Vec3 center;
-    double radius;
+    float radius;
     Vec3 rotation;
     shared_ptr<Material> mat;
 };
